+++ thor/Web/Form/Form.php
@@ -6,11 +6,11 @@
  * @license          MIT
  */
 
-namespace Thor\Html\Form;
+namespace Thor\Web\Form;
 
-use Thor\Html\Node;
+use Thor\Web\Node;
 use Thor\Http\Request\HttpMethod;
-use Thor\Html\Form\Field\FieldInterface;
+use Thor\Web\Form\Field\FieldInterface;
 
 abstract class Form extends Node implements FormInterface
 {
