+++ thor/Web/Node.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Thor\Html;
+namespace Thor\Web;
 
 class Node implements NodeInterface
 {
